Add Roboto as default font
This commit partially reverts commits b37279c8e and a255d1579. We're
moving the assets to the vendor folder, though, since it's where they
belong.

Note changing the font means selenium will double click on a space
character in a certain annotation test. So we're selecting the text
instead. The numbers "150" and "200" have been obtained by trial and
error.

Also note with the new font the columns selector "Price" element gets
out of the window when running the tests. While we investigate the
issue, we're just removing the part of the test dealing with the price.