Adapt custom verification residence to Santiago response format
As the Santiago response does not return the citizen date of birth nor
the postal code we cannot compare with data provided by the user so
we cannot grant that the user has the minimum age required to
participate.

Tests based on CensusApi were skipped (as this fork does not use it),
duplicated and adapted to pass when using the new remote census response
format, so we are sure we can still lock users after five failed attempts.