Add custom officing residence
As the web service response does not include the date_of_birth, we
are using the date provided through the interface. We still check if the
given date is above the minimum age, but it is not verified with the
web service so users could lie about its age.

Tests based on CensusApi were skipped (as this fork does not use it),
duplicated and adapted to pass when using the Santiago remote
census response format, so the officing residence features still work
even without the date_of_birth present on web service responses.